% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_MR_CLEAN}
\alias{df_MR_CLEAN}
\title{MR CLEAN trial data}
\format{
\subsection{\code{df_MR_CLEAN}}{

A data frame with 500 rows and 2 columns:
\describe{
\item{mRS}{Modified Rankin Scale}
\item{group}{Treatment group assignment; 1 = Intervention; 0 = Control}
\item{sex}{Sex indicator; 1 = women; 0 = men}
}
}
}
\source{
DOI: 10.1056/NEJMoa1411587
}
\usage{
df_MR_CLEAN
}
\description{
Patient outcome data from the MR CLEAN trial
}
\keyword{datasets}
