% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heat_box.R
\name{get_corr_heat_box}
\alias{get_corr_heat_box}
\title{Generate pairwise correlation heatbox}
\usage{
get_corr_heat_box(pairwise_cor_tbl, var1, var2, value)
}
\arguments{
\item{pairwise_cor_tbl}{a tbl containing three columns
\enumerate{
\item \code{column 1)} values to show on x-axis.
\item \code{column 2)} values to show on y-axis.
\item \code{column 2)} values denoting correlation between \code{column 1} and \code{column 2}.
}}

\item{var1}{a column to show on x-axis.}

\item{var2}{a column to show on y-axis.}

\item{value}{a column denoting corr values.}
}
\value{
an object of ggplot
}
\description{
Given a tibble of pairwise correlation get a correlation heatbox.
}
