% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_wise_scatter.R
\name{get_pair_wise_scatter}
\alias{get_pair_wise_scatter}
\title{Generate pairwise scatter plot}
\usage{
get_pair_wise_scatter(
  dat_tbl,
  group_tbl,
  var_plot,
  var_plot_group,
  dat_id,
  view_matrix = TRUE
)
}
\arguments{
\item{dat_tbl}{a tbl for which pairwise scatter plot to be generate.}

\item{group_tbl}{a tbl containing sample groups. Refer details for more information on groups.}

\item{var_plot}{a variable name, which is to be plotted.}

\item{var_plot_group}{a variable name, which is to be used to group the variable \code{var_plot}}

\item{dat_id}{a variable name from tbl \code{dat_tbl} storing feature ids. Typically name of first column form \code{dat_tbl}.}

\item{view_matrix}{logical, default TRUE, whether to display matrix view.}
}
\value{
an object of ggplot2
}
\description{
Generates pairwise scatter plot. One of the application of this function is to generates scatter plots between samples having 2 replicates.
}
