% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heat_box.R
\name{get_pairwise_cor_tbl}
\alias{get_pairwise_cor_tbl}
\title{Prepare pairwise corr tibble}
\usage{
get_pairwise_cor_tbl(dat, var = "genes", transform = TRUE, method = "pearson")
}
\arguments{
\item{dat}{a tbl.}

\item{var}{a column name denoting rownames. Default "genes".}

\item{transform}{logical, default TRUE, whether to log transform (log2) numeric columns.}

\item{method}{same as argument \code{method} in \code{stats::cor}.}
}
\value{
a tbl.
}
\description{
In general gene expression data are stored in tabular format, where first column is gene / feature name while rest other are expression values in respective samples/conditions.
Given such a data this function returns a long format tbl containing correlation value for each pair of variable.
}
