% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppairs.R
\name{ppairs}
\alias{ppairs}
\title{ppairs}
\usage{
ppairs(D, alph = 0.5)
}
\arguments{
\item{D}{data.frame or similar}

\item{alph}{transparency}
}
\value{
a ggplot2 object
}
\description{
A B&W smoothed contour corner plot
}
\details{
TODO
}
\author{
Pete Dodd
}
